fix: race condition in publisherToStream
A Stream created from a Publisher could end prematurely in the following scenario: Initially, no items are buffered. Downstream pulls. In `process`, we read the queue size and see it is 0. Then the current Fiber gets suspended. Meanwhile, the Publisher first send onNext and then signals onComplete. The new state is we have one item in the queue and the completion Promise is completed. Now the initial Fiber is resumed. It checks for completion without reading the queue size again, and ends the Stream, dropping the element in the queue.
This change masks the queue size check and the completion check as uninterruptible.