Channel that outputs to a subscriber.
Using a channel to drive a subscriber instead of a sink has the
advantage that we can pass upstream errors through to the
subscriber's `onError` method without needing the caller to
manually install an error handler.

While a channel is not as convenient to use as a sink, it is more
convenient that the error handler + sink pair.